pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "disasternet"
version = "1.0.0"
description = "Multi-Agency Disaster Response Coordination Environment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.2",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "numpy>=1.24.0",
    "huggingface-hub>=0.22.0",
]

[tool.setuptools]
packages = ["server"]
py-modules = ["models", "inference", "client", "ui"]

[project.scripts]
server = "server.app:main"